+++ src/os_api.h
@@ -24,7 +24,7 @@ typedef struct {
     // Cached allocations (handled by os)
     OS_Alloc *cache;
     void *sdl2_handle;
-    u64 uid;
+    u32 uid;
 } OS;
 
 // The only global variable allowed
